% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima_model}
\alias{arima_model}
\title{ARIMA model with fixed order}
\usage{
arima_model(p, d, q, intercept = TRUE, ...)
}
\arguments{
\item{p}{Order of auto-regressive (AR) terms.}

\item{d}{Degree of differencing}

\item{q}{Order of moving-average (MA) terms.}

\item{intercept}{Boolean value whether to include an intercept term (default: 
\code{TRUE}).}

\item{...}{Further arguments used when fitting ARIMA model.}
}
\value{
Model definition that can then be insered into \code{\link[caret]{train}}.
}
\description{
Creates an AIRMA model that is then fitted to the data as a univariate time series.
If further variables are specified in the model, it also includess exogenous variables. 
The order (p, d, q) is fixed as specified.
}
\note{
If one desires an auto-tuning of the best order, then one needs to switch to 
\code{\link{auto_arima_model}}.
}
\examples{
data(WWWusage) # from package "forecast"
df <- data.frame(y = as.numeric(WWWusage))

library(caret)

# without exogenous variables

lm <- train(y ~ 1, data = df, method = "lm", trControl = trainDirectFit())
summary(lm)

arima <- train(y ~ 1, data = df, method = arima_model(1, 1, 1), trControl = trainDirectFit())
summary(arima)
}

