% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{auto_arima_model}
\alias{auto_arima_model}
\title{ARIMA model with tuned order}
\usage{
auto_arima_model(p = 5, d = 2, q = 5, intercept = TRUE, ...)
}
\arguments{
\item{p}{Maximum order of auto-regressive (AR) terms that is tested to find the best 
fit (default: 5).}

\item{d}{Maximum degree of differencing that is tested to find the best fit (default: 
2).}

\item{q}{Maximum order of moving-average (MA) term that is tested to find the best 
fit (default: 5).}

\item{intercept}{Boolean value whether to include an intercept term (default: 
\code{TRUE}).}

\item{...}{Further arguments used when fitting ARIMA model.}
}
\value{
Model definition that can then be insered into \code{\link[caret]{train}}.
}
\description{
Creates an AIRMA model that is then fitted to the data as a univariate time series.
If further variables are specified in the model, it also includess exogenous variables. 
The order (p, d, q) is tuned by choosing the one with best fit.
}
\details{
Variable importance metrics return the absolute value of the coefficients
for the exogenous variables (if any).
}
\note{
If one desires an ARIMA model of fixed, pre-defined order, then one needs to 
switch to \code{\link{auto_arima_model}}.
}
\examples{
library(caret)

# without exogenous variables

library(forecast)
data(WWWusage) # from package "forecast"
df <- data.frame(y = as.numeric(WWWusage))

lm <- train(y ~ 1, data = df, method = "lm", trControl = trainDirectFit())
summary(lm)
RMSE(predict(lm, df), df)

arima <- train(y ~ 1, data = df, method = auto_arima_model(), trControl = trainDirectFit())
summary(arima)
RMSE(predict(arima, df), df)

# with exogenous variables

library(vars)
data(Canada)

arima <- train(x = Canada[, -2], y = Canada[, 2], 
               method = auto_arima_model(), trControl = trainDirectFit())

summary(arima)
arimaorder(arima$finalModel) # order of best model

predict(arima, Canada[, -2]) # in-sample predictions
RMSE(predict(arima, Canada[, -2]), Canada[, 2]) # in-sample RMSE

absCoef <- varImp(arima, scale = FALSE) # variable importance (= absolute value of coefficient)
absCoef

plot(absCoef)
}

