% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.R
\name{auto_arma_model}
\alias{auto_arma_model}
\title{ARMA model with tuned order}
\usage{
auto_arma_model(p = 5, q = 5, intercept = TRUE, ...)
}
\arguments{
\item{p}{Maximum order of auto-regressive (AR) terms that is tested to find the best 
fit (default: 5).}

\item{q}{Maximum order of moving-average (MA) term that is tested to find the best 
fit (default: 5).}

\item{intercept}{Boolean value whether to include an intercept term (default: 
\code{TRUE}).}

\item{...}{Further arguments used when fitting ARMA model.}
}
\value{
Model definition that can then be insered into \code{\link[caret]{train}}.
}
\description{
Creates an ARMA model that is then fitted to the data as a univariate time series.
If further variables are specified in the model, it also includess exogenous variables. 
The order (p, q) is tuned by choosing the one with best fit.
}
\details{
Variable importance metrics return the absolute value of the coefficients
for the exogenous variables (if any).
}
\note{
If one desires an ARMA model of fixed, pre-defined order, then one needs to 
switch to \code{\link{auto_arma_model}}.
}
\examples{
library(caret)

# without exogenous variables

library(forecast)
data(WWWusage) # from package "forecast"
df <- data.frame(y = as.numeric(WWWusage))

lm <- train(y ~ 1, data = df, method = "lm", trControl = trainDirectFit())
summary(lm)
RMSE(predict(lm, df), df)

arma <- train(y ~ 1, data = df, method = auto_arma_model(), trControl = trainDirectFit())
summary(arma)
RMSE(predict(arma, df), df)

# with exogenous variables

library(vars)
data(Canada)

arma <- train(x = Canada[, -2], y = Canada[, 2], 
              method = auto_arma_model(), trControl = trainDirectFit())

summary(arma)
arimaorder(arma$finalModel) # order of best model

predict(arma, Canada[, -2]) # in-sample predictions
RMSE(predict(arma, Canada[, -2]), Canada[, 2]) # in-sample RMSE

absCoef <- varImp(arma, scale = FALSE) # variable importance (= absolute value of coefficient)
absCoef

plot(absCoef)
}

