% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets.R
\name{ets_model}
\alias{ets_model}
\title{ETS model}
\usage{
ets_model(model = "ZZZ", ...)
}
\arguments{
\item{model}{Model type according to \code{\link[forecast]{ets}}. Default is 
\code{"ZZZ"} which performs auto-fitting.}

\item{...}{Further arguments used when fitting ETS model.}
}
\value{
Model definition that can then be insered into \code{\link[caret]{train}}.
}
\description{
Creates an exponential smoothing state space (ETS) model that is then fitted to the 
data as a univariate time series.
}
\note{
ETS model does not support exogenous variables.
}
\examples{
library(caret)

library(forecast)
data(WWWusage) # from package "forecast"

data_train <- WWWusage[1:80]
data_test <- WWWusage[81:100]

lm <- train(data_train, method = "lm", trControl = trainDirectFit())
summary(lm)
RMSE(predict(lm, numeric(20)), data_test)

ets <- train(data_train, method = ets_model(), trControl = trainDirectFit())
summary(ets)
RMSE(predict(ets, numeric(20)), data_test)
}

