% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train.numeric}
\alias{train.numeric}
\title{Train function to support single vector}
\usage{
\method{train}{numeric}(x, ...)
}
\arguments{
\item{x}{Object of class \code{numeric} in one dimension.}

\item{...}{Further arguments passed on to \code{\link[caret]{train}} for training.}
}
\value{
Trained model.
}
\description{
Extends the \code{\link[caret]{train}} function from the "caret" package to support
a single vector representing a time series.
}
\examples{
library(caret)

library(forecast)
data(WWWusage)

data_train <- WWWusage[1:80]
data_test <- WWWusage[81:100]

lm <- train(data_train, method = "lm", trControl = trainDirectFit())
summary(lm)
RMSE(predict(lm, data_test), data_test)
}

