% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train.ts}
\alias{train.ts}
\title{Train function to support time series objects}
\usage{
\method{train}{ts}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link[stats]{ts}}.}

\item{...}{Further arguments passed on to \code{\link[caret]{train}} for training.}
}
\value{
Trained model.
}
\description{
Extends the \code{\link[caret]{train}} function from the "caret" package to support
time series objects.
}
\examples{
library(caret)

library(forecast)
data(WWWusage)

class(WWWusage)
str(WWWusage)
arima <- train(WWWusage, method = auto_arima_model(), trControl = trainDirectFit())
summary(arima)
arimaorder(arima$finalModel) # order of best model
RMSE(predict(arima, df), df) # in-sample RMSE

library(vars)
data(Canada)

class(Canada)
str(Canada)

arima <- train(x = Canada[, -2], y = Canada[, 2], 
               method = auto_arima_model(), trControl = trainDirectFit())
summary(arima)
arimaorder(arima$finalModel) # order of best model
RMSE(predict(arima, df), df) # in-sample RMSE

arima <- train(form = prod ~ ., data = Canada, 
               method = auto_arima_model(), trControl = trainDirectFit())
summary(arima)
arimaorder(arima$finalModel) # order of best model
RMSE(predict(arima, df), df) # in-sample RMSE
}

