% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainControl.R
\name{trainDirectFit}
\alias{trainDirectFit}
\title{Training with direct fit only}
\usage{
trainDirectFit(...)
}
\arguments{
\item{...}{Optional parameters passed on to \code{\link[caret]{trainControl}}.}
}
\value{
An object that controls the training inside \code{\link[caret]{train}}.
}
\description{
This function is an alternative training function with performs a direct fit only,
i.e. it avoids any kind of cross-validation or any form of repeated executions. It
is thus beneficial to use this training function when working with objects that
solely require a direct fit, such as a least squares, ARMA or ARIMA. It thereby
considerably reduces runtime compare to the default initialization of 
\code{\link[caret]{trainControl}}.
}
\examples{
library(forecast)

data(WWWusage) # from package "forecast"
df <- data.frame(y = as.numeric(WWWusage))

library(caret)

lm <- train(y ~ 1, data = df, method = "lm", trControl = trainDirectFit())
summary(lm)
RMSE(predict(lm, df), df)

arima <- train(y ~ 1, data = df, method = arima_model(1, 1, 1), trControl = trainDirectFit())
summary(arima)
RMSE(predict(arima, df), df)
}

