% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima_model}
\alias{arima_model}
\title{ARIMA model with fixed order}
\usage{
arima_model(p, d, q, intercept = TRUE, ...)
}
\arguments{
\item{p}{Order of auto-regressive (AR) terms.}

\item{d}{Degree of differencing}

\item{q}{Order of moving-average (MA) terms.}

\item{intercept}{Boolean value whether to include an intercept term (default: 
\code{TRUE}).}

\item{...}{Further arguments used when fitting ARIMA model.}
}
\value{
Model definition that can then be insered into \code{\link[caret]{train}}.
}
\description{
Creates an AIRMA model that is then fitted to the data as a univariate time series.
If further variables are specified in the model, it also includess exogenous variables. 
The order (p, d, q) is fixed as specified.
}
\details{
Variable importance metrics return the absolute value of the coefficients
for the exogenous variables (if any).
}
\note{
If one desires an auto-tuning of the best order, then one needs to switch to 
\code{\link{auto_arima_model}}.
}
\examples{
library(caret)

# without exogenous variables

library(forecast)
data(WWWusage) # from package "forecast"
df <- data.frame(y = as.numeric(WWWusage))

lm <- train(y ~ 1, data = df, method = "lm", trControl = trainDirectFit())
summary(lm)

arima <- train(y ~ 1, data = df, method = arima_model(1, 1, 1), trControl = trainDirectFit())
summary(arima)

# with exogenous variables

library(vars)
data(Canada)

arima <- train(x = Canada[, -2], y = Canada[, 2], 
               method = arima_model(2, 0, 0), trControl = trainDirectFit())

summary(arima)
arimaorder(arima$finalModel) # order of best model

predict(arima, Canada[, -2]) # in-sample predictions
RMSE(predict(arima, Canada[, -2]), Canada[, 2]) # in-sample RMSE

absCoef <- varImp(arima, scale = FALSE) # variable importance (= absolute value of coefficient)
absCoef

plot(absCoef)
}

